requirements.txt
flask==2.0.1
flask-cors==3.0.10
requests==2.26.0
beautifulsoup4==4.9.3
python-docx==0.8.11
